+++ src/components/NavbarButtons.vue
@@ -11,7 +11,7 @@ const { isDarkTheme } = toRefs(styleStore);
     <c-button
       circle
       variant="text"
-      href="https://github.com/tbdavid2019/it-tools"
+      href="https://github.com/tbdavid2019/888tools"
       target="_blank"
       rel="noopener noreferrer"
       :aria-label="$t('home.nav.githubRepository')"