+++ src/stores/style.store.ts
@@ -1,25 +1,20 @@
 import { useDark, useMediaQuery, useStorage, useToggle } from '@vueuse/core';
 import { defineStore } from 'pinia';
-import { type Ref, watch } from 'vue';
 
-export const useStyleStore = defineStore('style', {
-  state: () => {
-    const isDarkTheme = useDark({ initialValue: 'light' });
-    const toggleDark = useToggle(isDarkTheme);
-    const isSmallScreen = useMediaQuery('(max-width: 700px)');
-    const isMenuCollapsed = useStorage('isMenuCollapsed', isSmallScreen.value) as Ref<boolean>;
-    const isBingWallpaperEnabled = useStorage('isBingWallpaperEnabled', true);
-    const cardOpacity = useStorage('cardOpacity', 0.8);
+export const useStyleStore = defineStore('style', () => {
+  const isDarkTheme = useDark({ initialValue: 'light' });
+  const toggleDark = useToggle(isDarkTheme);
+  const isSmallScreen = useMediaQuery('(max-width: 900px)');
+  const isMenuCollapsed = useStorage('isMenuCollapsed', isSmallScreen.value);
+  const isBingWallpaperEnabled = useStorage('isBingWallpaperEnabled', true);
+  const cardOpacity = useStorage('cardOpacity', 0.8);
 
-    watch(isSmallScreen, v => (isMenuCollapsed.value = v));
-
-    return {
-      isDarkTheme,
-      toggleDark,
-      isMenuCollapsed,
-      isSmallScreen,
-      isBingWallpaperEnabled,
-      cardOpacity,
-    };
-  },
+  return {
+    isDarkTheme,
+    toggleDark,
+    isMenuCollapsed,
+    isSmallScreen,
+    isBingWallpaperEnabled,
+    cardOpacity,
+  };
 });